Move funding finder to legacyPages, handle closed programmes
