+++ controllers/toplevel/legacyPages.js
@@ -1,12 +1,69 @@
+const Raven = require('raven');
 const config = require('config');
 const moment = require('moment');
 const ab = require('express-ab');
-const Raven = require('raven');
-const { get, includes } = require('lodash');
+const queryString = require('query-string');
+const { get, includes, toString } = require('lodash');
 const { legacyProxiedRoutes } = require('../routes');
 const { splitPercentages } = require('../../modules/ab');
 const { proxyLegacyPage, postToLegacyForm } = require('../../modules/proxy');
 const cached = require('../../middleware/cached');
+
+function reformatQueryString({ originalAreaQuery, originalAmountQuery }) {
+    originalAreaQuery = toString(originalAreaQuery).toLowerCase();
+    originalAmountQuery = toString(originalAmountQuery).toLowerCase();
+
+    let newQuery = {};
+    if (originalAreaQuery) {
+        newQuery.location = {
+            england: 'england',
+            'northern+ireland': 'northernIreland',
+            scotland: 'scotland',
+            wales: 'wales'
+        }[originalAreaQuery];
+    }
+
+    if (originalAmountQuery && originalAmountQuery !== 'up to 10000') {
+        newQuery.min = '10000';
+    }
+
+    return queryString.stringify(newQuery);
+}
+
+function normaliseQueryParam(req, name) {
+    // Old format URLs often get passed through as: ?area=Scotland&amp;amount=10001 - 50000
+    // urlencoded &amp; needs to be normalised when fetching individual query param
+    return req.query[`amp;${name}`] || req.query[name];
+}
+
+/**
+ * Legacy funding finder
+ * Proxy the legacy funding finder for closed programmes (where `sc` query is present)
+ * For all other requests normalise the query string and redirect to the new funding programmes list.
+ */
+function initLegacyFundingFinder(router) {
+    const fundingFinderPaths = [legacyProxiedRoutes.fundingFinder.path, legacyProxiedRoutes.fundingFinderWelsh.path];
+
+    fundingFinderPaths.forEach(mountPath => {
+        router
+            .get(mountPath, (req, res) => {
+                const showClosed = parseInt(normaliseQueryParam(req, 'sc'), 10) === 1;
+                if (showClosed) {
+                    // Proxy legacy funding finder for closed programmes
+                    return proxyLegacyPage(req, res, dom => dom, mountPath);
+                } else {
+                    // Redirect from funding finder to new programmes page
+                    const newQuery = reformatQueryString({
+                        originalAreaQuery: normaliseQueryParam(req, 'area'),
+                        originalAmountQuery: normaliseQueryParam(req, 'amount')
+                    });
+                    const redirectUrl = '/funding/programmes' + (newQuery.length > 0 ? `?${newQuery}` : '');
+                    res.redirect(301, redirectUrl);
+                }
+            })
+            .post(mountPath, postToLegacyForm);
+    });
+}
 
 function initAwardsForAll(router) {
     const awardsForAllRoutes = [
@@ -177,9 +234,11 @@ function initAwardsForAll(router) {
 }
 
 function init(router) {
+    initLegacyFundingFinder(router);
     initAwardsForAll(router);
 }
 
 module.exports = {
-    init
+    init,
+    reformatQueryString
 };
